requirements.txt
asgiref==3.5.2
colorama==0.4.5
Django==4.1.1
django-ckeditor==6.5.1
django-crispy-forms==1.14.0
django-js-asset==2.0.0
loguru==0.6.0
Pillow==9.2.0
sqlparse==0.4.2
tzdata==2022.2
win32-setctime==1.1.0
